+++ packages/axios-extension/src/abap/types/adt-types.ts
@@ -91,7 +91,13 @@ export type OperationsType = 'C' | 'P';
 export interface AtoSettings {
     developmentPackage?: string;
     developmentPrefix?: string;
+    /**
+     * Operations type cloud or on premise.
+     */
     operationsType?: OperationsType;
+    /**
+     * True if it is an S/4HANA Cloud Public Edition client for key user extensibility.
+     */
     isExtensibilityDevelopmentSystem?: boolean;
     tenantType?: TenantType;
     isTransportRequestRequired?: boolean;